+++ frontend/src/assets/adaptive-icon.png
